refactor: WorkPieceInformation 파일 분리
