+++ Expo1900/Expo1900/Model/Exhibition.swift
@@ -15,17 +15,4 @@ struct Exhibition: Decodable {
     let description: String
 }
 
-struct WorkPieceInformation: Decodable {
-    let name: String
-    let imageName: String
-    let shortDescription: String
-    let description: String
-    
-    enum CodingKeys: String, CodingKey {
-        case name
-        case imageName = "image_name"
-        case shortDescription = "short_desc"
-        case description
-    }
-}
